UPDATE | follow function

+++ src/main/java/com/weight/gym_dude/user/SiteUserImageRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SiteUserImageRepository extends JpaRepository<SiteUserImage, Long> {
-    Optional<SiteUserImage> findByAuthorId(Long id);
+public interface SiteUserImageRepository extends JpaRepository<SiteUserImage, Integer> {
+    Optional<SiteUserImage> findByAuthorId(Integer id);
 }
